requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pyyaml==6.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
pdfplumber==0.11.0
lxml==5.2.2
beautifulsoup4==4.12.3
javalang==0.13.0
tiktoken==0.7.0
aiofiles==24.1.0
